Recalculate grid size working (update vertex buffers at runtime)

+++ marching_cubes/distance_field.hpp
@@ -14,7 +14,7 @@
 
 
 // Create regular grid in the space of the point cloud
-std::vector<glm::vec3> create_regular_grid(int const& grid_resolution, std::vector<glm::vec3> const& point_cloud, std::vector<uint32_t>& );
+std::vector<glm::vec3> create_regular_grid(float const& grid_resolution, std::vector<glm::vec3> const& point_cloud, std::vector<uint32_t>& );
 
 // Indexes grid edges as a pair of vertices in the grid
 std::vector<std::pair<unsigned int, unsigned int>> index_grid_edges(std::vector<glm::vec3> const& grid_vertices);
